feat: automatically generate type definitions
