+++ package.json
@@ -66,7 +66,7 @@
     "sass-loader": "^11.1.0",
     "semver": "^6.2.0",
     "shelljs": "^0.8.4",
-    "typescript": "^4.0.3",
+    "typescript": "^4.3.3",
     "vue": "^3.1.1",
     "vue-analytics": "^5.16.1",
     "vue-loader": "^16.1.2",
