+++ packages/vuetify/.gitignore
@@ -5,4 +5,5 @@
 /es5/
 /lib/
 /lib-temp/
+/types-temp/
 /dist/
