+++ packages/vuetify/src/components/VDatePicker/util/eventHelpers.ts
@@ -1,9 +1,9 @@
 // @ts-nocheck
 /* eslint-disable */
 
-import Vue from 'vue'
+// import Vue from 'vue'
 
-export function createItemTypeNativeListeners (instance: Vue, itemTypeSuffix: string, value: any) {
+export function createItemTypeNativeListeners (instance: /*Vue*/any, itemTypeSuffix: string, value: any) {
   return Object.keys(instance.$listeners).reduce((on, eventName) => {
     if (eventName.endsWith(itemTypeSuffix)) {
       on[eventName.slice(0, -itemTypeSuffix.length)] = (event: Event) => instance.$emit(eventName, value, event)
@@ -13,7 +13,7 @@ export function createItemTypeNativeListeners (instance: Vue, itemTypeSuffix: st
   }, {} as typeof instance.$listeners)
 }
 
-export function createItemTypeListeners (instance: Vue, itemTypeSuffix: string) {
+export function createItemTypeListeners (instance: /*Vue*/any, itemTypeSuffix: string) {
   return Object.keys(instance.$listeners).reduce((on, eventName) => {
     if (eventName.endsWith(itemTypeSuffix)) {
       on[eventName] = instance.$listeners[eventName]
