+++ packages/vuetify/tsconfig.dist.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./tsconfig.json",
+  "exclude": ["dev", "**/*.spec.*", "cypress"]
+}
