+++ packages/vuetify/tsconfig.json
@@ -3,8 +3,10 @@
   "include": [
     "src"
   ],
+  "exclude": ["types-temp"],
   "compilerOptions": {
     "baseUrl": ".",
+    "outDir": "./types-temp",
     "paths": {
       "@/*": [
         "src/*"
