--- packages/vuetify/types/.npmignore
@@ -1,2 +1,0 @@
-# Ignore type tests
-test
