--- packages/vuetify/types/cypress.d.ts
@@ -1,9 +1,0 @@
-import type { mount as cyMount } from '@cypress/vue'
-
-declare global {
-  namespace Cypress {
-    export interface Chainable {
-      mount: typeof cyMount
-    }
-  }
-}
