--- packages/vuetify/types/services/goto.d.ts
@@ -1,31 +1,0 @@
-// @ts-nocheck
-// eslint-disable
-
-// Types
-import Vue from 'vue'
-
-type VuetifyGoToTarget = number | string | HTMLElement | Vue
-
-export type VuetifyGoToEasing =
-  ((t: number) => number) |
-  'linear' |
-  'easeInQuad' |
-  'easeOutQuad' |
-  'easeInOutQuad' |
-  'easeInCubic' |
-  'easeOutCubic' |
-  'easeInOutCubic' |
-  'easeInQuart' |
-  'easeOutQuart' |
-  'easeInOutQuart' |
-  'easeInQuint' |
-  'easeOutQuint' |
-  'easeInOutQuint'
-
-export interface GoToOptions {
-  container?: string | HTMLElement | Vue
-  duration?: number
-  offset?: number
-  easing?: VuetifyGoToEasing
-  appOffset?: boolean
-}
