--- packages/vuetify/types/test/lib.ts
@@ -1,38 +1,0 @@
-// @ts-nocheck
-// eslint-disable
-
-import { App, defineComponent } from 'vue'
-
-import Vuetify, {
-  useVuetify,
-  VBtn,
-  VCard,
-  VCardText,
-} from 'vuetify/lib'
-
-import * as directives from 'vuetify/lib/directives'
-
-declare const app: App
-
-app.use(Vuetify)
-app.use(Vuetify, {})
-app.use(Vuetify, {
-  components: {
-    VBtn,
-    VCard,
-    VCardText,
-  },
-  directives,
-})
-
-defineComponent({
-  setup () {
-    const vuetify = useVuetify()
-  },
-})
-
-VBtn.extend()
-
-Vue.extend({
-  extends: VBtn,
-})
