--- packages/vuetify/types/test/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "lib": [
-      "dom",
-      "es2015"
-    ],
-    "module": "commonjs",
-    "strict": true,
-    "noEmit": true,
-    "baseUrl": ".",
-    "paths": {
-      "vuetify": ["../../"]
-    }
-  },
-  "files": [
-    "../index.d.ts",
-    "../alacarte.d.ts",
-    "../lib.d.ts",
-    "alacarte.ts",
-    "global.ts",
-    "lib.ts",
-    "presets.ts"
-  ],
-  "compileOnSave": false
-}
